Workaround for openssl “symbol(s) not found”
During the build of the python dependency `openssl`, I was getting these
warnings:

    [DEBUG   ] clang: warning: using sysroot for 'MacOSX' but targeting 'iPhone' [-Wincompatible-sysroot]

and eventually the error

    [DEBUG   ] Undefined symbols for architecture x86_64:
    [DEBUG   ]   "_opendir$INODE64", referenced from:
    [DEBUG   ]       _OPENSSL_DIR_read in o_dir.o
    [DEBUG   ]   "_readdir$INODE64", referenced from:
    [DEBUG   ]       _OPENSSL_DIR_read in o_dir.o
    [DEBUG   ] ld: symbol(s) not found for architecture x86_64
    [DEBUG   ] clang: error: linker command failed with exit code 1 (use -v to see invocation)

kivy is trying to build OpenSSL for iOS using the macOS headers, then
failing at link time. The `--sysroot` / `-isysroot` compiler arguments
aren’t very well documented; [this test case] seems to indicate that
`--sysroot` is the main sysroot, and `-isysroot` is specific to headers? So
the original `--sysroot` might be preferred? Trying this in a clean VM
would help to sort it out.

[this test case]: https://github.com/llvm/llvm-project/blob/62ec4ac90738a5f2d209ed28c822223e58aaaeb7/clang/test/Driver/sysroot-flags.c